refactor(routing): FE-0 Refactors routing
Updates routing to allow JavaScript and TypeScript files to be used in routing.

Previously, we got route data by reading the file and then parsing the JSON.
This has been replaced with a dynamic import of the file, allowing us to better
rely on Webpack’s handling of file imports.